feat(bench): M5 — release mode, machine fingerprint, IQR gate
Implement ADR-0017 §"Run modes" release-mode rigor: 2 warmup + 10
measurement reps with a randomized cross-impl schedule (deterministic
given run_seed), CPU-governor pre-flight, IQR-relative-to-median gate
on the total stage at 5%, machine fingerprint in the result path, and
Linux-only platform gate.

- platform_check: Linux-only gate with the ADR's normative error message.
- machine: 12-hex sha256 fingerprint over (cpu_model, n_cores, total_ram,
  os_release, glibc); ensure_performance_governor() inspects every
  cpufreq/scaling_governor and emits a copy-pasteable cpupower fix.
- stats: per-stage median/IQR/min/max + iqr_gate(); release mode applies
  the gate to the total stage.
- orchestrate: MODE_REPS table (dev=0+1, release=2+10, profile=0+1);
  _build_schedule interleaves impls per rep so thermal drift can't
  systematically advantage one runner; tensor-of-record is rep 0 with
  per-rep bit-equality check; CellRun gains iqr_outcomes.
- cli: --mode {dev,release,profile} drives mode_defaults; profile auto-
  skips parity; ensure_linux() runs first so non-Linux fails cleanly.
- schema: IqrGateResult model + Aggregation.iqr_gate field.

Tests cover fingerprint stability across 100 invocations, schedule
determinism vs seed change, IQR math/threshold semantics, mode-default
table, Linux-only gate message, and governor pre-flight (mocking
_governor_paths so unit tests are filesystem-independent).

+++ bench/bench/workloads/jittered_predictions.py
@@ -82,7 +82,7 @@ def _generate(*, gt_path: Path, seed: int, out: Path) -> None:
             }
         )
 
-    n_fp = int(round(FP_FRACTION * len(non_crowd)))
+    n_fp = round(FP_FRACTION * len(non_crowd))
     if n_fp and image_size_by_id and gt.get("categories"):
         image_ids = list(image_size_by_id.keys())
         category_ids = [int(c["id"]) for c in gt["categories"]]
